Added code for new Alchemicals
Added code for Lithium, Honey, Chlorine and Stone Alchemicals. 
Lithium and Honey require Bunco and convert cards to Halberds and Fleurons.
Chlorine requires Reverie and can be used to instantly complete Cine quests.
Stone requires The Dread Jokers and is created when the Plague Doctor Joker rots an alchemical card.